// lib/api/model/courseOfferingStaffAccessLevel.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type CourseOfferingStaffAccessLevel = typeof CourseOfferingStaffAccessLevel[keyof typeof CourseOfferingStaffAccessLevel];


export const CourseOfferingStaffAccessLevel = {
  Owner: 'Owner',
  Instructor: 'Instructor',
  Assistant: 'Assistant',
  Viewer: 'Viewer',
} as const;