// lib/api/model/courseOfferingStaffScope.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type CourseOfferingStaffScope = typeof CourseOfferingStaffScope[keyof typeof CourseOfferingStaffScope];


export const CourseOfferingStaffScope = {
  Offering: 'Offering',
  Section: 'Section',
} as const;